feat: extract information from Sender
